+++ src/mako/database/query/compilers/MariaDB.php
@@ -68,7 +68,7 @@ class MariaDB extends MySQL
 	#[Override]
 	protected function vectorDistanceOrdering(array $order): string
 	{
-		return "{$this->vectorDistance($order)} {$order['order']}";
+		return "{$this->vectorDistance($order)} {$order['sortDirection']->value}";
 	}
 
 	/**