+++ src/mako/database/query/compilers/Oracle.php
@@ -82,7 +82,7 @@ class Oracle extends Compiler
 	{
 		$nullOrder = $order['nullsLast'] ? 'LAST' : 'FIRST';
 
-		return "{$this->column($order['column'])} {$order['order']} NULLS {$nullOrder}";
+		return "{$this->column($order['column'])} {$order['sortDirection']->value} NULLS {$nullOrder}";
 	}
 
 	/**